Add files via upload
Steganography is a technique used to hide secret information within non-secret data, making the existence of the hidden information undetectable. This can be done in various media such as images, audio, video files, or even text. The primary goal of steganography is to conceal the fact that a secret message exists, rather than just encrypting the content itself. It's widely used in digital watermarking, covert communications, and copyright protection.